splitting up index
